+++ bin/test.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-# Copyright 2023 Vivid Inc. and/or its affiliates.
+# Copyright 2024 Vivid Inc. and/or its affiliates.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.